refactor: reorganize components folder of yourLibrary
