+++ src/app/@yourLibrary/_components/Contents/AlbumItem.tsx
