+++ src/app/@yourLibrary/default.tsx
@@ -1,9 +1,20 @@
+import { auth } from '@/lib/utils/auth/auth'
 import Header from './_components/Header'
+import Contents from './_components/Contents'
+import CreatePlaylist from './_components/CreatePlaylist'
 
-const DefaultYourLibrary = () => {
+const DefaultYourLibrary = async () => {
+  const { isAuthenticated } = await auth()
   return (
-    <div className="h-full min-w-[250px] rounded-lg bg-gray-700">
+    <div className="flex h-full flex-col overflow-hidden rounded-lg bg-gray-700">
       <Header />
+      {isAuthenticated ? (
+        <div className="flex-1 overflow-y-scroll scrollbar-hide">
+          <Contents />
+        </div>
+      ) : (
+        <CreatePlaylist />
+      )}
     </div>
   )
 }
